+++ CHANGELOG.md
@@ -20,6 +20,7 @@ CHANGELOG
 * Updated captions/subtiles file fetching to support cross-origin requests in older IE browsers [[view](https://github.com/videojs/video.js/pull/1095)]
 * Added support for playback rate switching [[view](https://github.com/videojs/video.js/pull/1132)]
 * Fixed an issue with the loadstart event order that caused the big play button to not hide [[view](https://github.com/videojs/video.js/pull/1209)]
+* Modernized the fullscreen API and added support for IE11 [[view](https://github.com/videojs/video.js/pull/1205)]
 
 --------------------